Conseguido: siguiente y anterior, cargar de bbdd una vez, insertar emociones, ensenar feliz

+++ js/emusic.js
@@ -50,21 +50,24 @@ module.controller('MyCtrl', function($scope, $cordovaMedia) {
   media.getCurrentPosition().then(...);
 });
 */
+
+//PONER UN POPUP DE CARGANDO MIENTRAS SE CARGAN LAS CANCIONES (para un futuro):
+/*
+$('#waiting-popup').popup('open');
+$('#waiting-popup').popup('close');*/
+
 var nombre;
-/*---------------------Añadido Sabadoo--------------*/
-var listaCanciones;
-/*--------------------------------------------------------*/
+var listaCanciones = [];
+
 $(document).on( "pagecreate", "#player-page", function( e ) {
-    var titulo = (($(this).data("url").indexOf("?") > 0) ? $(this).data("url") : '-' ).replace( /.*titulo=/, "" ).replace(new RegExp("\\+","g"),' ');
+    var titulo = (($(this).data("url").indexOf("?") > 0) ? $(this).data("url") : '-' ).replace(new RegExp("\\+","g"),' ');
     nombre = titulo.split('=')[1];
     $("#media-name").text(nombre);//¿¿¿¿HACER EN EL SETINTERVAL PARA CUANDO SE PASE DE CANCIÓN SE ACTUALICE EL NOMBRE?????
 
     hacerSelectCancion(nombre, iniciarAudio);
 
-    /*---------------------Añadido Sabadoo--------------*/
-    listaDeCanciones(nombre);//Carga la lista de canciones en una variable global para ser usada cuando se quiera pasar de canción
+    listaDeCanciones();//Carga la lista de canciones en una variable global para ser usada cuando se quiera pasar de canción
 
-    /*--------------------------------------------------------*/
     // Pause after 10 seconds
     /*setTimeout(function () {
         my_media.pause();
@@ -74,6 +77,7 @@ $(document).on( "pagecreate", "#player-page", function( e ) {
 
 window.setInterval(function(){
 	if(my_media){
+		$("#media-name").text(nombre);
 		$('#media-duration').html("<span>"+Math.round(((my_media.getDuration())/60)*100)/100+"</span>");//Durancion en minutos redondeada a dos decimales
 		var pos;
 		my_media.getCurrentPosition(function(position){
@@ -92,15 +96,18 @@ function play() {
 function playMusic(nombre) {
 	if (status == 2) {//está reproduciendo
 		hacerSelectCancion(nombre, pausaAudio);
-	} else if (status == 3 || status == 4) {//está en pausa
+	} else if (status == 3 || status == 4) {//está en pausa o stop
 		hacerSelectCancion(nombre, playAudio);
 	}
 }
 var iniciarAudio = function(url) {
+	if(my_media){
+		my_media.stop()
+	}
     // Play the audio file at url
     my_media = new Media(url,
         // success callback
-        function () { alert("playAudio():Audio Success"); },
+        function () { /*alert("playAudio():Audio Success");*/ },
         // error callback
         function (err) { alert("playAudio():Audio Error: " + err); },
         function(mediaStatus) {
@@ -108,7 +115,7 @@ var iniciarAudio = function(url) {
         }
     );
 
-
+    estamosPrimeraCarga = false;
     // Play audio
     my_media.play();
 }
@@ -150,17 +157,6 @@ function refrescarCanciones() {
 	$(":mobile-pagecontainer").pagecontainer("change", "todas_canciones.html", {transition: "pop"});
 }
 
-/*function buscarAudio()
-{
-	window.resolveLocalFileSystemURL(cordova.file.dataDirectory, function(dirEntry) {
-		console.log('file sistem abierto: ' + dirEntry.name);
-
-		var para = document.createElement("P");
-		var v = document.createTextNode(dirEntry.name);
-		para.appendChild(v);
-		document.getElementById('files-list').appendChild(para);
-	}, onErrorLoadFs);
-}*/
 var getRootDir = function(entry) {
 	entry.getParent(function(padre){
 		padre.getParent(function(padre){
@@ -179,10 +175,11 @@ function hacerSelectCancion(nombre, callback) {
 	});
 }
 
+/*DEBUGEO BORRAAAAAAAAAAAAAAAAAAAAAAAAAAAAR*/
 function hacerSelect() {
 	var result = [];
 	db.transaction(function (tx) {
-	  	tx.executeSql('SELECT * FROM TODAS_MUSICA', [], function(tx, rs){
+	  	tx.executeSql('SELECT * FROM FELIZ', [], function(tx, rs){
 
 	    for(var i=0; i<rs.rows.length; i++) {
 	    	var row = rs.rows.item(i);
@@ -220,12 +217,16 @@ function buscardirectorio (entry, nivel){
 }
 
 var db;
+var borrarTablaMusica = true;
 function guardarEnBD(nombreCancion, direccionCancion){
 	db = window.openDatabase("emusic", "1.0", "Cordova Demo", 200000);
 	db.transaction(insertarCancion, errorCB, successCB);
 
 	function insertarCancion(tx) {
-		tx.executeSql('DROP TABLE IF EXISTS TODAS_MUSICA');
+		if (borrarTablaMusica) {
+			tx.executeSql('DROP TABLE IF EXISTS TODAS_MUSICA');
+			borrarTablaMusica = false;
+		}
 	    tx.executeSql('CREATE TABLE IF NOT EXISTS TODAS_MUSICA (nombre varchar(30), path varchar(255) NOT NULL, PRIMARY KEY (path))');
 		tx.executeSql('INSERT INTO TODAS_MUSICA (nombre, path) VALUES (?, ?)', [nombreCancion, direccionCancion]);
 		$(function() {
@@ -242,35 +243,55 @@ function guardarEnBD(nombreCancion, direccionCancion){
 	    //alert("success!");
 	}
 }
-function cargarBBDD() {
+/*function cargarBBDD() {
 	window.resolveLocalFileSystemURI("file:///canciones.txt", function (fs) {
     	fs.root.getFile("canciones.txt", { create: true, exclusive: false }, function (fileEntry) {
 			readFile(fileEntry);
     	}, onErrorCreateFile);
 	}, onErrorLoadFs);
-}
+}*/
+
+var estamosPrimeraCarga = true;
 
 $(document).on( "pagecontainerchange",function(){
 	var pageID = $(':mobile-pagecontainer').pagecontainer('getActivePage')[0].id;
 	if(pageID == "files-list-page"){
-		var permissions = cordova.plugins.permissions;
-		permissions.requestPermission(permissions.READ_EXTERNAL_STORAGE, successR, errorR);
-		function errorR() {
-		  alert("No tenemos permisos");
-		}
-		function successR( status ) {
-		  if( !status.hasPermission ) error();
-		}
+		if (estamosPrimeraCarga) {
+			borrarTablaMusica = true;
+			var permissions = cordova.plugins.permissions;
+			permissions.requestPermission(permissions.READ_EXTERNAL_STORAGE, successR, errorR);
+			function errorR() {
+			  alert("No tenemos permisos");
+			}
+			function successR( status ) {
+			  if( !status.hasPermission ) error();
+			}
 
-		permissions.requestPermission(permissions.WRITE_EXTERNAL_STORAGE, successW, errorW);
-		function errorW() {
-		  alert("No tenemos permisos");
-		}
-		function successW( status ) {
-		  if( !status.hasPermission ) error();
-		}
+			permissions.requestPermission(permissions.WRITE_EXTERNAL_STORAGE, successW, errorW);
+			function errorW() {
+			  alert("No tenemos permisos");
+			}
+			function successW( status ) {
+			  if( !status.hasPermission ) error();
+			}
 
-		window.resolveLocalFileSystemURL(cordova.file.externalApplicationStorageDirectory, getRootDir);//ACCEDE A LA CARPETA DE LA APLICACIÓN
+			window.resolveLocalFileSystemURL(cordova.file.externalApplicationStorageDirectory, getRootDir);//ACCEDE A LA CARPETA DE LA APLICACIÓN
+		} else {
+			$(function() {
+				for (var i = 0; i < listaCanciones.length; i++) {
+					$("#files-list").append('<li><a href="reproductor.html?nombre='+listaCanciones[i].nombre+'" data-transition="flip" data-role="button">'+listaCanciones[i].nombre+'</a></li>').listview('refresh');
+				}
+			});
+		}
+	}
+	else if(pageID == "page_feliz"){
+		cargarListaFeliz();
+	}
+	else if(pageID == "page_serio"){
+		cargarListaSerio();
+	}
+	else if(pageID == "page_triste"){
+		cargarListaSerio();
 	}
 });
 
@@ -294,10 +315,9 @@ $(document).on( "pagecontainerchange",function(){
 }*/
 
 
-function listaDeCanciones(nombre){//Casi igual a hacerSelect(), para usar hacerSelect tenia que sacar el array fuera y no quería tocarlo, en caso de que se pueda sacar, sacarlo y usar hacerSelect en vez de esta función
+var listaDeCanciones = function() {//Casi igual a hacerSelect(), para usar hacerSelect tenia que sacar el array fuera y no quería tocarlo, en caso de que se pueda sacar, sacarlo y usar hacerSelect en vez de esta función
 	db.transaction(function(tx){
 		tx.executeSql('SELECT * FROM TODAS_MUSICA ORDER BY nombre', [], function(tx, rs){
-
 		    for(var i=0; i<rs.rows.length; i++) {
 		    	var row = rs.rows.item(i);
 		        listaCanciones[i] = {nombre: row['nombre'], fullPath: row['path']};
@@ -307,45 +327,57 @@ function listaDeCanciones(nombre){//Casi igual a hacerSelect(), para usar hacerS
 
 	});
 
-}*
+}
 function pasarCancion(siguienteAnterior){//0-> siguiente, 1-> anterior Se llama desde los onclick de los botones prev y next del reporductor
 	if(siguienteAnterior==0){
-		siguienteCancion(nombre,listaCanciones);
+		siguienteCancion();
 	}
 	else if(siguienteAnterior==1){
-		anteriorCancion(nombre,listaCanciones);
+		anteriorCancion();
 	}
 }
-function siguienteCancion(nombre,listaCanciones){
-	for(var i=0; i<listaCanciones.length;i++){
+function siguienteCancion(){
+	var encontrada = false;
+	var i=0;
+	while (!encontrada && i<listaCanciones.length) {
 		if(listaCanciones[i].nombre==nombre){
-			callback(playAudio(listaCanciones[i+1].url));//Reproduce la siguiente canción de la lista, la lista esta ordenada alfabeticamente
+			encontrada = true;
+			nombre = listaCanciones[(i+1)%listaCanciones.length].nombre;
+			iniciarAudio(listaCanciones[(i+1)%listaCanciones.length].fullPath);//Reproduce la siguiente canción de la lista, la lista esta ordenada alfabeticamente
 		}
+		i++;
 	}
-
 }
-function anteriorCancion(nombre,listaCanciones){
-	for(var i=0; i<listaCanciones.length;i++){
+function anteriorCancion(){
+	var encontrada = false;
+	var i=0;
+	while (!encontrada && i<listaCanciones.length) {
 		if(listaCanciones[i].nombre==nombre){
-			callback(playAudio(listaCanciones[i-1].url));//Reproduce la anterior canción de la lista, la lista esta ordenada alfabeticamente
+			encontrada = true;
+			if (i==0) {
+				nombre = listaCanciones[listaCanciones.length-1].nombre;
+				iniciarAudio(listaCanciones[listaCanciones.length-1].fullPath);//Reproduce la siguiente canción de la lista, la lista esta ordenada alfabeticamente
+			} else {
+				nombre = listaCanciones[(i-1)%listaCanciones.length].nombre;
+				iniciarAudio(listaCanciones[(i-1)%listaCanciones.length].fullPath);//Reproduce la siguiente canción de la lista, la lista esta ordenada alfabeticamente
+			}
 		}
+		i++;
 	}
 }
 /*---------------------------Insertar canción a lista de reproducción de emociones---------------------------------*/
 
 function DistinguirEmocion(emocion){//0->Feliz, 1->Triste, 2->Serio
-	if(emocion==0){
-		AñadirCancionEmocion(nombre,AñadirFeliz);
-	}
-	else if(emocion==1){
-		AñadirCancionEmocion(nombre,AñadirTriste);
-	}
-	else if(emocion==2){
-		AñadirCancionEmocion(nombre,AñadirSerio);
+	if (emocion==0) {
+		anadirCancionEmocion(nombre,anadirFeliz);
+	} else if (emocion==1) {
+		anadirCancionEmocion(nombre,anadirTriste);
+	} else if (emocion==2) {
+		anadirCancionEmocion(nombre,anadirSerio);
 	}
 }
 
-function AñadirCancionEmocion(nombre,callback){//Funcion que busca el path de la canción y llama a la función de añadir canción de la emoción que nos indiquen
+function anadirCancionEmocion(nombre,callback){//Funcion que busca el path de la canción y llama a la función de añadir canción de la emoción que nos indiquen
 	db.transaction(function (tx) {
 		tx.executeSql('SELECT path FROM TODAS_MUSICA WHERE nombre=?', [nombre], function(tx, rs){
 			callback(nombre, rs.rows.item(0)['path']);
@@ -353,7 +385,7 @@ function AñadirCancionEmocion(nombre,callback){//Funcion que busca el path de l
 	});
 }
 
-var AñadirFeliz=function (nombre,path){
+var anadirFeliz=function (nombre,path){
 	db.transaction(insertarCancion, errorCB, successCB);
 
 	function insertarCancion(tx) {
@@ -361,14 +393,14 @@ var AñadirFeliz=function (nombre,path){
 		tx.executeSql('INSERT INTO FELIZ (nombre, path) VALUES (?, ?)', [nombre, path]);
 	}
 	function errorCB(err) {
-	    alert("Error processing SQL: "+err.code);
+	    alert("¡Esta canción ya ha sido añadida a la lista FELIZ!");
 	}
 	function successCB() {
 	    //alert("success!");
 	}
 }
 
-var AñadirTriste=function (nombre,path){
+var anadirTriste=function (nombre,path){
 	db.transaction(insertarCancion, errorCB, successCB);
 
 	function insertarCancion(tx) {
@@ -376,14 +408,14 @@ var AñadirTriste=function (nombre,path){
 		tx.executeSql('INSERT INTO TRISTE (nombre, path) VALUES (?, ?)', [nombre, path]);
 	}
 	function errorCB(err) {
-	    alert("Error processing SQL: "+err.code);
+	    alert("¡Esta canción ya ha sido añadida a la lista TRISTE!");
 	}
 	function successCB() {
 	    //alert("success!");
 	}
 }
 
-var AñadirSerio=function (nombre,path){
+var anadirSerio=function (nombre,path){
 	db.transaction(insertarCancion, errorCB, successCB);
 
 	function insertarCancion(tx) {
@@ -391,7 +423,7 @@ var AñadirSerio=function (nombre,path){
 		tx.executeSql('INSERT INTO SERIO (nombre, path) VALUES (?, ?)', [nombre, path]);
 	}
 	function errorCB(err) {
-	    alert("Error processing SQL: "+err.code);
+	    alert("¡Esta canción ya ha sido añadida a la lista SERIO!");
 	}
 	function successCB() {
 	    //alert("success!");
@@ -399,7 +431,7 @@ var AñadirSerio=function (nombre,path){
 }
 
 /*--------------------------Mostrar canciones del sentimiento---------------------------------------------*/
-//No se muy bien como llamarlas, pues no comprendo como hemos llamado a guardarEndb para que se carge cuando esta en todas_canciones.html
+
 function cargarListaFeliz(){
 	db.transaction(cargarLista, errorCB, successCB);
 	function cargarLista(tx) {
@@ -413,7 +445,7 @@ function cargarListaFeliz(){
 					//Aunque le pasemos en href 'reproductor', pues cuando vaya a reproducir no mirar de donde viene sino el parametro que se le pase
 				});
 			}
-		  }, null);
+		}, null);
 	}
 	function errorCB(err) {
 	    alert("Error processing SQL: "+err.code);
@@ -421,9 +453,8 @@ function cargarListaFeliz(){
 	function successCB() {
 	    //alert("success!");
 	}
-
 }
-
+/*
 function cargarListaTriste(){
 	db.transaction(cargarLista, errorCB, successCB);
 	function cargarLista(tx) {
@@ -470,8 +501,8 @@ function cargarListaSerio(){
 	    //alert("success!");
 	}
 
-}
-
+}*/
+/*------------------------------------------------------------------------------------*/
 /*
 function readFile(fileEntry) {
